Address comments in PR review
* Move EB config to config.global.yml
* Add migrations back in
* Run collectstatic normally
* Increase memory limit on app container (fixes disk i/o issue)
* Remove .elasticbeanstalk from .gitignore
* Use pip for supervisor
* Revert fabconfig changes
