+++ .ebextensions/01migrate.config
@@ -1,0 +1,4 @@
+container_commands:
+  01migrate:
+    command: "mkdir -p /opt/elasticbeanstalk/hooks/appdeploy/post/ && echo -e '#!/bin/bash\ndocker exec `docker ps -a -q | head -n 1` /usr/local/bin/entrypoint.sh setupproddb' > /opt/elasticbeanstalk/hooks/appdeploy/post/99_migrate.sh && chmod +x /opt/elasticbeanstalk/hooks/appdeploy/post/99_migrate.sh"
+    leader_only: true
