+++ .gitignore
@@ -90,7 +90,6 @@ celerybeat-schedule
 *Dockerrun.aws.json
 conf/nginx.ureport.conf
 conf/deploy.zip
-*.elasticbeanstalk
 
 # Elastic Beanstalk Files
 .elasticbeanstalk/*
