+++ Dockerfile
@@ -2,7 +2,6 @@ FROM ubuntu:14.04
 
 ENV HOME /root
 ENV TERM screen-256color
-ENV DEBIAN_FRONTEND noninteractive
 
 RUN locale-gen --no-purge en_US.UTF-8
 ENV LC_ALL en_US.UTF-8
@@ -32,13 +31,16 @@ ADD conf/nginx.ureport.conf /etc/nginx/sites-enabled/default
 WORKDIR /code
 RUN mkdir -p /var/www/static && chmod -R 760 /var/www/static/ && chown -R www-data:www-data /var/www/static
 
+RUN mkdir -p /etc/supervisor/conf.d && mkdir -p /var/log/supervisor
+
 ADD ./ /code/
 
+RUN ln -sf /code/conf/supervisord.conf /etc/supervisord.conf
 RUN ln -sf /code/conf/supervisor.ureport.conf /etc/supervisor/conf.d/ureport.conf
 RUN ln -sf /code/conf/supervisor.nginx.conf /etc/supervisor/conf.d/nginx.conf
 RUN ln -sf /usr/bin/nodejs /usr/bin/node
 
-EXPOSE 80
+EXPOSE 80 8000
 
 ENTRYPOINT ["/usr/local/bin/entrypoint.sh"]
 CMD ["start"]
