+++ conf/supervisor.ureport.conf
@@ -5,3 +5,4 @@ autostart=true
 autorestart=true
 stdout_logfile=/var/log/ureport/ureport.log
 redirect_stderr=True
+environment=RDS_HOSTNAME="%(ENV_RDS_HOSTNAME)s",RDS_USERNAME="%(ENV_RDS_USERNAME)s",RDS_PASSWORD="%(ENV_RDS_PASSWORD)s",RDS_PORT="%(ENV_RDS_PORT)s",RDS_DB_NAME="%(ENV_RDS_DB_NAME)s",REDISHOST="%(ENV_REDISHOST)s",STATIC_DIR="%(ENV_STATIC_DIR)s"
