+++ pip-requires.txt
@@ -1,0 +1,35 @@
+-e git+https://github.com/nyaruka/smartmin.git#egg=smartmin
+Django==1.8.4
+-e git+https://github.com/rapidpro/rapidpro-python.git@master#egg=rapidpro-python-2.0
+-e git+https://github.com/rapidpro/dash.git#egg=dash
+celery
+coverage
+django-celery
+django-compressor
+django-debug-toolbar
+django-digest
+django-guardian
+django-nose
+django-reversion
+django-rosetta
+django-storages
+django-rest-swagger
+djangorestframework
+dj-database-url
+boto
+Pillow
+celery-with-redis
+gunicorn
+nose
+pisa
+raven
+django-redis
+psycopg2
+xlutils
+mock
+sorl-thumbnail==11.12.1b
+hamlpy
+pycountry
+django-countries
+GitPython
+regex
