+++ ureport/settings.py.dev
@@ -5,3 +5,5 @@ INSTALLED_APPS = INSTALLED_APPS + ('django_nose',)
 
 DEBUG = True
 TEMPLATE_DEBUG = DEBUG
+
+COMPRESS_OFFLINE = False
